requirements.txt
pip==19.3.1
django
python-telegram-bot[socks]
Pillow
pre-commit
django-mysql
requests
bs4
lxml
